requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
streamlit
shap
xgboost
plotly
joblib
umap-learn
imblearn